+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+* Added the *--skip-pack-dependencies* flag to **validate** command to skip pack dependencies validation. Relevant when using the *-g* flag.
 
 # 1.1.9
 * Fixed an issue where **update-release-notes** command failed on invalid file types.